backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
pymysql==1.1.0
python-decouple==3.8
pdfplumber==0.11.8
PyPDF2==3.0.1

